Created additional package for user dtos
