+++ src/main/java/pl/sggw/wzimlibrary/controller/RegistrationController.java
@@ -8,7 +8,7 @@ import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 import pl.sggw.wzimlibrary.exception.UserAlreadyExistsException;
 import pl.sggw.wzimlibrary.exception.UserDecryptionException;
 import pl.sggw.wzimlibrary.model.User;
-import pl.sggw.wzimlibrary.model.dto.UserRegistrationDto;
+import pl.sggw.wzimlibrary.model.dto.user.UserRegistrationDto;
 import pl.sggw.wzimlibrary.service.RegistrationService;
 import pl.sggw.wzimlibrary.service.UserService;
