+++ src/main/java/pl/sggw/wzimlibrary/model/dto/user/UserChangePasswordVerificationDto.java
@@ -1,22 +1,22 @@
-package pl.sggw.wzimlibrary.model.dto;
+package pl.sggw.wzimlibrary.model.dto.user;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import pl.sggw.wzimlibrary.model.constant.SecurityQuestion;
 import pl.sggw.wzimlibrary.model.constant.ValidationConstant;
 
+import javax.validation.constraints.Email;
 import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
 
 @RequiredArgsConstructor
 @Getter
-public class UserPanelChangeQuestionDto {
+public class UserChangePasswordVerificationDto {
 
     @NotEmpty
-    @Size(min = ValidationConstant.PASSWORD_MIN_LENGTH,
-            max = ValidationConstant.PASSWORD_MAX_LENGTH)
-    private final String password;
+    @Email
+    private final String email;
 
     @NotNull
     private final SecurityQuestion securityQuestion;
@@ -24,5 +24,4 @@ public class UserPanelChangeQuestionDto {
     @NotEmpty
     @Size(max = ValidationConstant.SECURITY_QUESTION_ANSWER_MAX_LENGTH)
     private final String securityQuestionAnswer;
-
 }
